# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eBPFsentinel Agent API",
    "description": "REST API for the eBPFsentinel network security agent.\n\nProvides endpoints for managing firewall rules, IPS/IDS policies, rate limiting, L7 filtering, threat intelligence, alerts, and audit logs.\n\nMost endpoints require authentication via JWT Bearer token or API key.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "v0.0.0-dev"
  },
  "paths": {
...
    "/api/v1/routing/gateways": {
      "get": {
        "tags": [
          "Routing"
        ],
        "summary": "`GET /api/v1/routing/gateways` - list routing gateways.",
        "operationId": "list_gateways",
        "responses": {
          "200": {
            "description": "Gateway list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GatewayResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Routing"
        ],
        "summary": "`POST /api/v1/routing/gateways` - add a routing gateway.",
        "operationId": "create_gateway",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGatewayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Gateway created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Gateway conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/routing/gateways/{id}": {
      "delete": {
        "tags": [
          "Routing"
        ],
        "summary": "`DELETE /api/v1/routing/gateways/{id}` - remove a routing gateway.",
        "operationId": "delete_gateway",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Gateway identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Gateway deleted"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Gateway not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/routing/routes": {
      "get": {
        "tags": [
          "Routing"
        ],
        "summary": "`GET /api/v1/routing/routes` - effective default route(s) from gateway selection.",
        "operationId": "list_routes",
        "responses": {
          "200": {
            "description": "Active routes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/routing/status": {
      "get": {
        "tags": [
          "Routing"
        ],
        "summary": "`GET /api/v1/routing/status` - routing status.",
        "operationId": "routing_status",
        "responses": {
          "200": {
            "description": "Routing status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutingStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateGatewayRequest": {
        "type": "object",
        "required": [
          "name",
          "ip"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "health_check_interval_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Health-check probe interval in seconds. When set, an ICMP probe is configured.",
            "minimum": 0
          },
          "interface": {
            "type": [
              "string",
              "null"
            ],
            "description": "Egress interface; defaults to empty (inherits the agent's primary)."
          },
          "ip": {
            "type": "string",
            "description": "Gateway IPv4 address."
          },
          "name": {
            "type": "string"
          },
          "weight": {
            "type": "integer",
            "format": "int32",
            "description": "Routing weight; maps to failover priority (lower = preferred).",
            "minimum": 0
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `RULE_NOT_FOUND`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error."
          }
        }
      },
...
      "GatewayResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "interface",
          "gateway_ip",
          "priority",
          "weight",
          "enabled",
          "status",
          "health_status"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "gateway_ip": {
            "type": "string"
          },
          "health_status": {
            "type": "string",
            "description": "Alias of `status` for clients expecting a `health_status` field."
          },
          "id": {
            "type": "string",
            "description": "Gateway identifier, rendered as a string for stable API addressing."
          },
          "interface": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "description": "Failover priority (lower = preferred).",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "Health-check observed status (`healthy` / `degraded` / `down`)."
          },
          "weight": {
            "type": "integer",
            "format": "int32",
            "description": "Alias of `priority`, exposed as a routing weight.",
            "minimum": 0
          }
        }
      },
...
      "RouteResponse": {
        "type": "object",
        "description": "A resolved default route derived from the active gateway selection.",
        "required": [
          "destination",
          "gateway_id",
          "gateway_ip"
        ],
        "properties": {
          "destination": {
            "type": "string",
            "description": "Destination CIDR served by this route."
          },
          "gateway_id": {
            "type": "string",
            "description": "Identifier of the gateway this route egresses through."
          },
          "gateway_ip": {
            "type": "string"
          }
        }
      },
      "RoutingStatusResponse": {
        "type": "object",
        "required": [
          "enabled",
          "gateway_count"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "gateway_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key configured in agent"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained via OIDC or static configuration"
      }
    }
  },
...
}